feat(parsing): model token mismatches as integer penalties
Replace the float multiplier config with integer 0-100 penalties converted
to multipliers at scoring time, and partially penalize a year present on
only one side. Rename match_weights config to token_weights/token_multipliers.

+++ src/subsearch/runtime/models/model.py
@@ -52,7 +52,7 @@ class AppConfig:
     only_foreign_parts: bool
     providers: dict[str, bool]
     token_weights: dict[str, float]
-    token_multipliers: dict[str, float]
+    token_multipliers: dict[str, int]
     context_menu: bool
     context_menu_icon: bool
     file_extensions: dict[str, bool]
